change bash

+++ bash.sh
@@ -5,15 +5,25 @@
 
 case $system in
   ($linux)
-    if $CUSTOM_BASHRC > /dev/null 2>&1; then
+    if [ ! -f ~/.bash ]; then
       print_command \
       "add custom '.bashrc'"
 
-      echo source ~/setting/bashrc >> ~/.bashrc
+      echo source ~/.bash >> ~/.bashrc
+      ln bashrc ~/.bash
     fi
     ;;
 
   ($mac)
-    ln bashrc ~/.bash_profile
+    if [ ! -f ~/.bash_profile ]; then
+      print_command \
+      "add custom '.bashrc'"
+
+      ln bashrc ~/.bash_profile
+    fi
     ;;
 esac
+
+if [ ! -e ~/Desktop/work ]; then
+  mkdir ~/Desktop/work
+fi
